refactor: gather code imports that do not rely on composer autoloading
Centralize the import of the app global constants
along with that of the composer default autoloader.
It is all declared in a single script which is imported in the front
controller
of the app. The composer default autoloader used to be imported right
there.
The imports of the app global constants are removed from where they used
to be